orchestrator/requirements.txt
a2a-sdk>=0.2.0
claude-agent-sdk>=0.1.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
starlette>=0.40.0
python-dotenv>=1.0.0
